[DEV-3687] updating redux
