+++ src/js/redux/reducers/index.js
@@ -19,7 +19,6 @@ import downloadReducer from './search/downloadReducer';
 import bulkDownloadReducer from './bulkDownload/bulkDownloadReducer';
 import redirectModalReducer from './redirectModal/redirectModalReducer';
 import stateReducer from './state/stateReducer';
-import awardV2Reducer from './awardV2/awardReducer';
 
 const appReducer = combineReducers({
     filters: filtersReducer,
@@ -35,8 +34,7 @@ const appReducer = combineReducers({
     explorer: explorerReducer,
     bulkDownload: bulkDownloadReducer,
     redirectModal: redirectModalReducer,
-    stateProfile: stateReducer,
-    awardV2: awardV2Reducer
+    stateProfile: stateReducer
 });
 
 export default appReducer;
